bimtester: move code into a package directory
