+++ src/ifcbimtester/bimtester/clean.py
