+++ src/ifcbimtester/bimtester/guiwidget.py
@@ -13,11 +13,15 @@ from .run import run_all
 class GuiWidgetBimTester(QtWidgets.QWidget):
 
     # get some initial values
+    this_path = os.path.join(os.path.dirname(__file__))
+    desktop_path = os.path.join(os.path.expanduser("~"), "Desktop")
+
     initial_ifcfile = "/home/hugo/Documents/zeug_sort/z_some_ifc/example_model.ifc"
     if not os.path.isfile(initial_ifcfile):
-        initial_ifcfile = os.path.join(os.path.expanduser("~"), "Desktop")
-    this_path = os.path.join(os.path.dirname(__file__))
-    initial_featurespath = os.path.join(this_path, "..", "features_bimtester", "fea_min")
+        initial_ifcfile = desktop_path
+    initial_featurespath = os.path.join(this_path, "..", "..", "features_bimtester", "fea_min")
+    if not os.path.isdir(initial_featurespath):
+        initial_featurespath = desktop_path
 
     def __init__(self):
         super(GuiWidgetBimTester, self).__init__()
