+++ src/ifcbimtester/bimtester/resources/icons/bimtester.ico
